+++ FirstProject/Controllers/FeaturesSectionsController.cs
@@ -8,6 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using FirstProject.Models;
 using AspNetCoreHero.ToastNotification.Abstractions;
 using Microsoft.AspNetCore.Http;
+using FirstProject.Infrastructure;
 
 namespace FirstProject.Controllers
 {
@@ -28,7 +29,7 @@ namespace FirstProject.Controllers
             var checkAdmin = HttpContext.Session.GetInt32("AdminId");
             if (checkAdmin != null)
             {
-                var AdminInformation = await _context.Admins.FindAsync((decimal)checkAdmin);
+                var AdminInformation = await _context.Admins.FindAsync(checkAdmin);
                 ViewBag.AdminId = checkAdmin;
                 ViewBag.AdminName = AdminInformation.FirstName;
                 ViewBag.AdminIamge = AdminInformation.ImagePath;
@@ -39,7 +40,7 @@ namespace FirstProject.Controllers
         }
 
         // GET: FeaturesSections/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -58,12 +59,12 @@ namespace FirstProject.Controllers
 
         
         // GET: FeaturesSections/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             var checkAdmin = HttpContext.Session.GetInt32("AdminId");
             if (checkAdmin != null)
             {
-                var AdminInformation = await _context.Admins.FindAsync((decimal)checkAdmin);
+                var AdminInformation = await _context.Admins.FindAsync(checkAdmin);
                 ViewBag.AdminId = checkAdmin;
                 ViewBag.AdminName = AdminInformation.FirstName;
                 ViewBag.AdminIamge = AdminInformation.ImagePath;
@@ -88,7 +89,7 @@ namespace FirstProject.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("Id,TimesunWed,TimethuFri,TimesatSun,EmaegencyCase")] FeaturesSection featuresSection)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,TimesunWed,TimethuFri,TimesatSun,EmaegencyCase")] FeaturesSection featuresSection)
         {
             if (id != featuresSection.Id)
             {
@@ -119,7 +120,7 @@ namespace FirstProject.Controllers
             return View(featuresSection);
         }
 
-        private bool FeaturesSectionExists(decimal id)
+        private bool FeaturesSectionExists(int id)
         {
             return _context.FeaturesSections.Any(e => e.Id == id);
         }
